Refactor the AllReader class to have separate readers for different csvs. Added a sample for postpage csv. Modified the test cases and some javadoc.
