+++ src/main/interfaceadaptor/contollers/Controller.java
@@ -1,0 +1,4 @@
+package interfaceadaptor.contollers;
+
+public class Controller {
+}
