+++ src/main/interfaceadaptor/readers/AllReader.java
@@ -1,9 +1,10 @@
-package interfaceadaptor;
+package interfaceadaptor.readers;
 
 import entity.Professor;
 import entity.Student;
-import usecase.createupdate.CourseCreator;
-import usecase.enrolldropcourse.CourseEnroller;
+import interfaceadaptor.CsvReader;
+import usecase.creator.ProfessorCreator;
+import usecase.creator.StudentCreator;
 import usecase.javastorage.AllProfessors;
 import usecase.javastorage.AllStudents;
 
@@ -13,7 +14,6 @@ import java.util.Objects;
 public class AllReader {
     public static void readAll(){
         ArrayList<ArrayList<String>> users = CsvReader.readCsv("/Users/hanqizhang/Desktop/CSC207/course-project-utalk11/src/main/outerlayer/database/user.csv");
-        ArrayList<ArrayList<String>> courses = CsvReader.readCsv("/Users/hanqizhang/Desktop/CSC207/course-project-utalk11/src/main/outerlayer/database/courses.csv");
         ArrayList<ArrayList<String>> students = CsvReader.readCsv("/Users/hanqizhang/Desktop/CSC207/course-project-utalk11/src/main/outerlayer/database/students.csv");
         ArrayList<ArrayList<String>> professors = CsvReader.readCsv("/Users/hanqizhang/Desktop/CSC207/course-project-utalk11/src/main/outerlayer/database/professors.csv");
         for (ArrayList<String> user: users){
@@ -25,22 +25,9 @@ public class AllReader {
                 AllProfessors.ProfessorHashMap.put(user.get(1), professor);
             }
         }
-        for (ArrayList<String> course: courses){
-            CourseCreator.createCourse(course.get(0), course.get(1), course.get(2));
-        }
-        for (ArrayList<String> student: students){
-            String studentName = student.get(0);
-            student.remove(0);
-            for (String course:student){
-                CourseEnroller.enrollCourse(course, AllStudents.StundetHashMap.get(studentName));
-            }
-        }
-        for (ArrayList<String> professor: professors){
-            String professorName = professor.get(0);
-            professor.remove(0);
-            for (String course: professor){
-                CourseEnroller.teachingCourse(course, AllProfessors.ProfessorHashMap.get(professorName));
-            }
-        }
+        CourseReader.readCourse();
+        StudentCreator.createStudent(students);
+        ProfessorCreator.createProfessor(professors);
+
     }
 }
